requirements.txt
requests
Pillow
# torch
pandas